+++ src/web/mod.rs
@@ -1,2 +1,5 @@
+pub mod abort;
+pub mod base64;
 pub mod blob;
+pub mod performance;
 pub mod url_search_params;